requirements.txt
appdirs==1.4.3
elasticsearch==2.3.0
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
oauthlib==2.0.1
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
requests==2.13.0
requests-aws4auth==0.9
requests-oauthlib==0.8.0
six==1.10.0
tweepy==3.5.0
urllib3==1.20
Werkzeug==0.12
